Implement the Multiple Choice Form Item
This form item is similar to the Single Choice Form item, except that it allows for more than one selection. The ID values that are sent to the server are formatted in a comma-separated fashion.

The selected values are shown as a RecyclerView, styled as Android Chips, horizontally distributed and above the input Spinner.

On the other hand, the displaying and editing of an existing workflow with selected values for the Multiple Choice Form Item is pending and will be developed in a further commit.