Implement the Phone Form Item
The phone form item is similar to the currency form item, in the way that it is divided into two inputs, a country spinner and a text input. Those values are serialized into a JSON object that is sent to the server.

The filling and editing of this form item is still pending.