Create the AutocompleteFormItem
This item is used for those SingleSelection fields that have way too many options, and instead of presenting a list of options, the user needs to type a search query first, then select one of the searched options.

As of now, this item is used by the Accounts field type.